nvidia: merge all test cases into a single module
